Move PaginationComponent component to components folder
