+++ src/components/PaginationComponent/styles.js
